@@ -2,13 +2,14 @@ package main
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"log"
-	"math/rand"
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
+)
+
+const (
+	serverPort = 8585
 )
 
 func main() {
@@ -18,7 +19,7 @@ func main() {
 	}
 
 	// Loading game assets.
-	b, err := ioutil.ReadFile(os.Args[1])
+	b, err := os.ReadFile(os.Args[1])
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -27,21 +28,17 @@ func main() {
 		log.Fatalf("corrupted assets file: %s\n", err)
 	}
 
-	// Initializing the RNG.
-	rand.Seed(time.Now().UnixNano())
-
 	// Capture OS signals.
 	sig := make(chan os.Signal, 1)
 	signal.Notify(sig, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
 
 	// Run server.
-	srv := newServer(8585)
+	srv := newServer(serverPort)
 	srv.start(sig)
 
 	// Initializing the game.
 	//g := newGame(a)
 	//
-	//// --- DEBUG ---
 	//fmt.Println(g.Companies)
 	//fmt.Println(g.ActionDeck)
 	//fmt.Println(g.BankDeck)
